+++ src/main/java/com/teamtea/eclipticseasons/common/core/biome/WeatherManager.java
@@ -48,6 +48,7 @@ import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.common.util.FakePlayer;
 import net.minecraftforge.common.util.INBTSerializable;
 import net.minecraftforge.fml.loading.FMLLoader;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.*;
 
@@ -64,6 +65,22 @@ public class WeatherManager {
             return null;
         }
         return BIOME_WEATHER_LIST.getOrDefault(level, null);
+    }
+
+    public static BiomeWeather getBiomeWeather(Level level, Holder<Biome> biomeHolder) {
+        if (biomeHolder == null) return null;
+        return getBiomeWeather(level, biomeHolder.value());
+    }
+
+    public static BiomeWeather getBiomeWeather(Level level, Biome biome) {
+        ArrayList<BiomeWeather> ws = getBiomeList(level);
+        if (ws != null)
+            for (BiomeWeather biomeWeather : ws) {
+                if (biome == biomeWeather.biomeHolder.value()) {
+                    return biomeWeather;
+                }
+            }
+        return null;
     }
 
     public static Level fetchLevelIfNull(Level level) {
@@ -587,6 +604,13 @@ public class WeatherManager {
             );
 
         }
+    }
+
+    public static int getSkyDarken(Level level, BlockPos pPos) {
+        WeatherManager.BiomeWeather biomeWeather = WeatherManager.getBiomeWeather(level, MapChecker.getSurfaceBiome(level, pPos));
+        int amount = biomeWeather == null || biomeWeather.shouldClear() ? 0 :
+                biomeWeather.shouldThunder() ? 8 : 4;
+        return Mth.clamp(amount, 0, 15);
     }