+++ src/main/java/com/teamtea/eclipticseasons/mixin/common/MixinLevelReader.java
@@ -11,16 +11,22 @@ import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.biome.Biome;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Overwrite;
 
 @Mixin({LevelReader.class})
 public interface MixinLevelReader extends LevelReader {
 
+    /**
+     * @author jianzoushihu ( joe vettek)
+     * @reason Ecliptic Seasons adjusts the weather system to be localized under Solar Weather conditions,
+     * requiring brightness information to be corrected based on block positions.
+     */
+    @Overwrite
     @Override
     default int getMaxLocalRawBrightness(@NotNull BlockPos pPos) {
         int amount = this.getSkyDarken();
         if (this instanceof Level level && EclipticUtil.hasLocalWeather(level)) {
-            amount += WeatherManager.getRainOrSnow(level, MapChecker.getSurfaceBiome(level,pPos).value(), pPos) != Biome.Precipitation.NONE ? 10 : 0;
-            amount = Mth.clamp(amount, 0, 15);
+            amount = WeatherManager.getSkyDarken(level, pPos);
         }
         return this.getMaxLocalRawBrightness(pPos, amount);
     }